Refactor code, add dao layer, InvestmentService class and implement functionality to save Investment entity to db, add in memory db test, update pom.xml with spring test dependency
Refactor code, add dao layer, InvestmentService class and implement functionality to save Investment entity to db, add in memory db test, update pom.xml with spring test dependency

Refactor code, add dao layer, InvestmentService class and implement functionality to save Investment entity to db, add in memory db test, update pom.xml with spring test dependency# You are currently rebasing branch 'main' on 'a3278c6'.

+++ src/main/java/com/enjoythecode/bitcoininvestmentsimulator/model/Investment.java
@@ -42,11 +42,11 @@ public class Investment implements Serializable {
     private BigDecimal profitAmount;
 
     @Column(precision = 19, scale = 2)
-    private BigDecimal profitPercentage;
+    private double profitPercentage;
 
     public Investment(LocalDate calculationDate, LocalDate investmentDate, BigDecimal investmentAmount, String currency,
                       BigDecimal btcAmount, BigDecimal btcPriceOnInvestmentDate, BigDecimal btcPriceToday,
-                      BigDecimal profitAmount, BigDecimal profitPercentage) {
+                      BigDecimal profitAmount, double profitPercentage) {
         this.calculationDate = calculationDate;
         this.investmentDate = investmentDate;
         this.investmentAmount = investmentAmount;
